% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopNull.R
\name{stopNull}
\alias{stopNull}
\title{stopNull}
\usage{
stopNull(except = NULL)
}
\arguments{
\item{except}{args' name vector that should not to be checked}
}
\description{
Check caller's args and stop if there are any NULLs.
}
\examples{
add = function(a, b, c) {
	stopNull()
	a + b + c
}

add(1, 2, 3)

add(1, 2, NULL)

add2 = function(a, b, c) {
	stopNull(except = "c")
	a + b
}

add2(1, 2, NULL)
}
